+++ src/img2txt/characters/braillesegment.py
