+++ src/img2txt/colors/colordifference.py
